Add async-grpo rollout trace logging to trackio
